+++ src/snippets/validators/go.rs
@@ -207,6 +207,7 @@ mod tests {
                 "GOCACHE".into(),
                 root.path().join("go-cache").to_string_lossy().into_owned(),
             )]),
+            include_paths: Vec::new(),
             rust_features: Vec::new(),
             rust_dependencies: BTreeMap::new(),
         };